moved mockpipelinedb to its own file
