+++ apps/passport-server/src/services/genericIssuanceService.ts
@@ -234,32 +234,6 @@ export interface PipelineAtomDB {
 }
 
 /**
- * A mock implementation of {@link PipelineAtomDB} for testing purposes.
- */
-export class MockPipelineAtomDB {
-  public data: {
-    [pipelineId: string]: { [atomId: string]: PipelineAtom };
-  } = {};
-
-  public async save(pipelineID: string, atoms: PipelineAtom[]): Promise<void> {
-    if (!this.data[pipelineID]) {
-      this.data[pipelineID] = {};
-    }
-    atoms.forEach((atom) => {
-      this.data[pipelineID][atom.id] = atom;
-    });
-  }
-
-  public async load(pipelineID: string): Promise<PipelineAtom[]> {
-    if (!this.data[pipelineID]) {
-      return [];
-    }
-
-    return Object.values(this.data[pipelineID]);
-  }
-}
-
-/**
  * {@link Pipeline}s can have external-facing APIs. Anything that an external
  * user can do with a Pipeline needs to be represented as a {@link PipelineCapability}.
  */
